# Improve block propagation performance